Make convertEventHandlers private
